Migrate all API routes from Pages Router to App Router
- Migrate 6 routes: contact, subscribe, unsubscribe, increment-view,
  list-view-count, og-image to app/api/*/route.ts
- Create lib/route-helpers.ts with checkRateLimit and checkCsrf guards
  using Web API (NextRequest/NextResponse) instead of NextApiRequest
- Remove pages/ directory entirely (no more Pages Router)
- Remove old lib/rate-limit.ts and lib/csrf-protection.ts
